+++ doc/sphinx/changelog.rst
@@ -9,7 +9,7 @@ Changelog
 New
 ~~~
 
-- Add a :cpp:func:`pagmo::base_bgl_topology::get_weight()`
+- Add a :cpp:func:`pagmo::base_bgl_topology::get_edge_weight()`
   function to fetch the weight of an edge in a BGL topology
   (`#407 <https://github.com/esa/pagmo2/pull/407>`__).